REF: simplify should_extension_dispatch, remove dispatch_to_extension_op
